* Halo1_BlamLib: build-cache-file rewrite progress * Halo1_BlamLib: Fixed cache_files_windows.cpp not compiling * Halo1_BlamLib: changed scenario's netgame_equipment block field to be declared non-const in runtime * Halo1_CheApe: s_main_globals can now be accessed

+++ OpenSauce/shared/Include/blamlib/Halo1/main/levels.hpp
@@ -41,7 +41,7 @@ namespace Yelo
 
 	namespace blam
 	{
-		int32 PLATFORM_API main_get_campaign_level_from_name(cstring level_name);
+		int16 PLATFORM_API main_get_campaign_level_from_name(cstring level_name);
 
 		cstring PLATFORM_API main_get_campaign_level_name(_enum level_index);
 	};
